VerifyBot/requirements.txt
nextcord==2.6.0
python-dotenv==1.0.1
Pillow==10.4.0
captcha==0.5.0
aiohttp==3.10.5